requirements.txt
psycopg2==2.9.1
pysam==0.19.1
PyYAML>=6.0
